# openapi/libraryz.yaml
openapi: 3.1.0
info:
  title: LibraryZ API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    LibraryZ backend: auth and sessions, catalog (works + editions with
    content-addressed file storage), crowdsourced edits with moderation,
    personal libraries, and recommendations. See PLAN.md and ARCHITECTURE.md
    in the repo root.

    **Every operation has a cURL sample** (`x-codeSamples`). They assume:

    ```sh
    BASE=http://localhost:8080          # Kong in `docker compose up`
    TOKEN=...      REFRESH=...          # from POST /auth/login (see its sample)
    WORK_ID=...    EDITION_ID=...    CONTRIBUTION_ID=...
    ```

    **Authentication.** `POST /auth/login` returns a short-lived access token
    (JWT, 15 minutes by default) and a single-use refresh token. Send the
    access token as `Authorization: Bearer <token>`. On a `401`, call
    `POST /auth/refresh` once and retry; if that also fails, log in again.
    Presenting an already-used refresh token revokes the whole session.

    **Errors.** `4xx`/`5xx` bodies are short `text/plain` messages (`4xx`
    messages are safe to show to users), except where an operation documents
    a JSON body (e.g. the `409` on edition upload). `5xx` bodies are generic;
    details are only logged server-side.

    **Rate limits.** In the compose/production setup Kong enforces per-IP
    limits (login 5/min, signup 3/min, refresh/logout 20/min, edition upload
...
servers:
...
    description: Local dev
...
security: []
paths:
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/auth/signup" -H 'Content-Type: application/json' \
              -d '{"email":"you@example.com","password":"hunter22","name":"You"}'
      responses:
        "201":
          description: Created
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "409":
          description: Email already registered
          content:
            text/plain:
              schema: { type: string }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Exchange credentials for an access + refresh token pair
      description: |
        Starts a new session. The access token is a short-lived JWT
        (`LIBRARYZ_ACCESS_TOKEN_TTL`, default 15m); renew it with
        `POST /auth/refresh`. The access token is also mirrored in the
        `Authorization` response header.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            # Keep both tokens: the access token expires in 15 minutes.
            curl -s -X POST "$BASE/auth/login" -H 'Content-Type: application/json' \
              -d '{"email":"you@example.com","password":"hunter22"}' > tokens.json
            TOKEN=$(jq -r .access_token tokens.json)
            REFRESH=$(jq -r .refresh_token tokens.json)
      responses:
        "200":
          description: Authenticated
          headers:
            Authorization:
              description: "`Bearer <access_token>`"
              schema:
                type: string
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [auth]
      summary: Rotate a refresh token for a new token pair
      description: |
        The presented refresh token is spent and a new pair is returned.
        Presenting an already-spent token is treated as theft: the whole
        session (token family) is revoked and the client must log in again.
        Clients must therefore serialize refreshes per session.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            # Single use: store the new refresh_token, the old one is now spent.
            curl -s -X POST "$BASE/auth/refresh" -H 'Content-Type: application/json' \
              -d "{\"refresh_token\":\"$REFRESH\"}" > tokens.json
            TOKEN=$(jq -r .access_token tokens.json)
            REFRESH=$(jq -r .refresh_token tokens.json)
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Unknown, expired, revoked, or reused refresh token — log in again
          content:
            text/plain:
              schema: { type: string }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the session a refresh token belongs to
      description: |
        Revokes the refresh token's session. No access token needed (it may
        already have expired). Unknown tokens are ignored. Outstanding access
        tokens for the session stay valid until they expire (≤ access TTL);
        use `/auth/logout-all` to kill those immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/auth/logout" -H 'Content-Type: application/json' \
              -d "{\"refresh_token\":\"$REFRESH\"}"
      responses:
        "204":
          description: Logged out
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/logout-all:
    post:
      operationId: logoutAll
      tags: [auth]
      summary: End every session of the current user
      description: |
        Increments the user's token version — invalidating every outstanding
        access token at once — and revokes all refresh tokens.
      security:
        - bearerAuth: []
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/auth/logout-all" -H "Authorization: Bearer $TOKEN"
      responses:
        "204":
          description: All sessions ended
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Return the currently authenticated user
      description: |
        Reads live DB state for `is_moderator`, not the JWT — so a user
        promoted via SQL while logged in will see the new flag without
        needing to re-login.
      security:
        - bearerAuth: []
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/auth/me" -H "Authorization: Bearer $TOKEN"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /works:
    get:
      operationId: listWorks
      tags: [catalog]
      summary: List works
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/works?limit=50&offset=0"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Work" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      operationId: createWork
      tags: [catalog]
      summary: Create a new work
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkRequest"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/works" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"title":"Moby-Dick","authors":"Herman Melville","publication_year":1851}'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Work" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /works/search:
    get:
      operationId: searchWorks
      tags: [catalog]
      summary: Search works by free-text query
      description: |
        Matches against `title`, `subtitle`, `authors`, and `description`,
        most relevant first. On Postgres uses a weighted `tsvector` GIN index
        (`simple` configuration — no stemming / stopwords) queried with
        `websearch_to_tsquery`, so `"exact phrase"`, `or`, and `-exclude`
        work; results are ordered by `ts_rank` (title > subtitle/authors >
        description), then `created_at DESC`. On sqlite (the test DB) falls
        back to a literal, case-insensitive substring match (`%` and `_` are
        not wildcards) ordered title hits, then subtitle/author hits, then
        description hits, then `created_at DESC`. Empty `q` returns 400.
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -G "$BASE/works/search" \
              --data-urlencode 'q="herman melville" -abridged' \
              --data-urlencode limit=20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Work" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /works/{id}:
    get:
      operationId: getWork
      tags: [catalog]
      summary: Get a single work with its editions and tags
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/works/$WORK_ID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Work" }
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      operationId: removeWork
      tags: [moderation]
      summary: Remove a work (moderator, or its creator while empty)
      description: |
        Moderators can take down any work; any other user can remove only a
        work they created that has no editions (the upload client does this
        when a new work's first upload is rejected). Removes the work and all
        of its editions. Soft
        delete: the rows are kept with who/when/why for audit (and can be
        restored in the DB), but the work disappears from list, search, get,
        download, personal libraries and recommendations. The files are
        purged by the blob GC after `LIBRARYZ_BLOB_GC_RETENTION`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RemovalRequest" }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X DELETE "$BASE/works/$WORK_ID" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"reason":"DMCA notice #42"}'
      responses:
        "204":
          description: Removed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /works/{id}/editions:
    post:
      operationId: uploadEdition
      tags: [editions]
      summary: Upload a new edition (file) for a work
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [format, file]
              properties:
                format:
                  type: string
                  description: "pdf | epub | mobi | azw3 | djvu | cbz | txt | ..."
                language:
                  type: string
                  description: BCP-47 language tag (optional)
                file:
                  type: string
                  format: binary
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/works/$WORK_ID/editions" -H "Authorization: Bearer $TOKEN" \
              -F format=epub -F language=en -F file=@moby-dick.epub
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Edition" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            The file is already in the library as an edition, on this work or
            another one. Matched on the sha256 of the uploaded bytes, or of the
            bytes after sanitization. Nothing is recorded. If that edition was
            removed by a moderator, `edition_id` / `work_id` are omitted and
            `error` says so (a takedown can't be undone by re-uploading).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DuplicateEdition" }
        "413":
          description: >
            The file couldn't be sanitized within the server's memory or time
            budget (for example, a PDF whose compressed object streams
            inflate to gigabytes). Nothing is recorded.
          content:
            text/plain:
              schema: { type: string }
        "415":
          description: Format is not on the allowlist (currently PDF, EPUB, TXT)
          content:
            text/plain:
              schema: { type: string }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /editions/{id}:
    get:
      operationId: getEdition
      tags: [editions]
      summary: Get edition metadata
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/editions/$EDITION_ID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Edition" }
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      operationId: removeEdition
      tags: [moderation]
      summary: Remove an edition (moderator)
      description: |
        Moderator-only takedown of one edition (soft delete; see
        `DELETE /works/{id}`). Re-uploading byte-identical content afterwards
        is refused with 409.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RemovalRequest" }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X DELETE "$BASE/editions/$EDITION_ID" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"reason":"broken file"}'
      responses:
        "204":
          description: Removed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /works/{id}/contributions:
    post:
      operationId: submitContribution
      tags: [contributions]
      summary: Submit a proposed edit to a work
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContributionRequest"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/works/$WORK_ID/contributions" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"patch":{"title":"Moby-Dick; or, The Whale","publication_year":1851}}'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contribution" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /contributions:
    get:
      operationId: listContributions
      tags: [contributions]
      summary: List contributions
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/contributions?status=pending&limit=50"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Contribution" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /contributions/{id}:
    get:
      operationId: getContribution
      tags: [contributions]
      summary: Get a single contribution
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/contributions/$CONTRIBUTION_ID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contribution" }
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /contributions/{id}/approve:
    post:
      operationId: approveContribution
      tags: [contributions]
      summary: Approve a pending contribution and apply its patch to the work
      description: |
        Moderator-only. Atomically moves the contribution out of `pending`
        and applies its `patch` to the target Work in the same transaction;
        if another moderator decided it first, returns 409 and applies
        nothing.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/contributions/$CONTRIBUTION_ID/approve" -H "Authorization: Bearer $TOKEN"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contribution" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Contribution is not pending
          content:
            text/plain:
              schema: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /contributions/{id}/reject:
    post:
      operationId: rejectContribution
      tags: [contributions]
      summary: Reject a pending contribution (no changes applied to the work)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/contributions/$CONTRIBUTION_ID/reject" -H "Authorization: Bearer $TOKEN"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contribution" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Contribution is not pending
          content:
            text/plain:
              schema: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /me/contributions:
    get:
      operationId: listMyContributions
      tags: [contributions]
      summary: List the calling user's own contributions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/me/contributions?status=pending" -H "Authorization: Bearer $TOKEN"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Contribution" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/library:
    get:
      operationId: listLibrary
      tags: [library]
      summary: List the calling user's personal-library entries
      description: |
        Each entry embeds the target `work` (with its editions) so the My
        Library screen renders without a second round-trip. Ordered by
        `updated_at` descending.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [want, reading, read] }
        - in: query
          name: shelf
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/me/library?status=reading" -H "Authorization: Bearer $TOKEN"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserBook" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /me/recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: Personalized work recommendations for the calling user
      description: |
        Naive v0: content-based on shared authors/tags with the user's liked
        library items (rating-weighted), with a popularity fallback for cold
        start / short result sets. Read-only; never recommends a work already
        in the user's library.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/me/recommendations?limit=20" -H "Authorization: Bearer $TOKEN"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recommendation" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/recommendations/{id}/dismiss:
    post:
      operationId: dismissRecommendation
      tags: [recommendations]
      summary: Hide a work from the caller's future recommendations
      description: "`{id}` is the work id. Idempotent."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/me/recommendations/$WORK_ID/dismiss" -H "Authorization: Bearer $TOKEN"
      responses:
        "204":
          description: Dismissed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token from `POST /auth/login` or `POST /auth/refresh`. Expires
        after `LIBRARYZ_ACCESS_TOKEN_TTL` (15m); revoked immediately by
        `POST /auth/logout-all`.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        text/plain:
          schema: { type: string }
...
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema: { type: string }
...
    InternalError:
      description: Server error
      content:
        text/plain:
          schema: { type: string }
  schemas:
...
    SignUpRequest:
      type: object
      required: [email, password, name]
      properties:
        email:    { type: string, format: email }
        password: { type: string, minLength: 1 }
        name:     { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string }
...